diploma-verifier/requirements.txt
# Framework web
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# OCR
pytesseract==0.3.10

# PDF
PyMuPDF==1.24.3
pdf2image==1.17.0

# Vision / Image
opencv-python-headless==4.9.0.80
scikit-image==0.23.2
Pillow==10.3.0
numpy==1.26.4

# NLP
spacy==3.7.5
langdetect==1.0.9

# Validation
pydantic==2.7.1

# Utilitaires
python-magic-bin==0.4.14
aiofiles==23.2.1

# Tests
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
